refactor and review ok
